Added OIDC based authorization feature

+++ lib/Controller/OpenProjectAPIController.php
@@ -86,6 +86,20 @@ class OpenProjectAPIController extends Controller {
 		$this->logger = $logger;
 	}
 
+	private function validatePreRequestConditions(): ?DataResponse {
+		$authMethod = $this->config->getAppValue(Application::APP_ID, 'authorization_method', '');
+		if ($authMethod === OpenProjectAPIService::AUTH_METHOD_OAUTH && $this->accessToken === '') {
+			return new DataResponse('', Http::STATUS_UNAUTHORIZED);
+		} elseif ($authMethod === OpenProjectAPIService::AUTH_METHOD_OIDC &&
+			$this->openprojectAPIService->getOIDCBasedTokenForTheTargetedAudienceClient($this->config->getAppValue(Application::APP_ID, 'targeted_audience_client_id', '') !== null)
+		) {
+			return new DataResponse('', Http::STATUS_UNAUTHORIZED);
+		} elseif (!OpenProjectAPIService::validateURL($this->openprojectUrl)) {
+			return new DataResponse('', Http::STATUS_BAD_REQUEST);
+		}
+		return null;
+	}
+
 	/**
 	 * get openproject instance URL
 	 * @NoAdminRequired
@@ -95,7 +109,6 @@ class OpenProjectAPIController extends Controller {
 	public function getOpenProjectUrl(): DataResponse {
 		return new DataResponse($this->openprojectUrl);
 	}
-
 	/**
 	 * get openproject user avatar
 	 * @NoAdminRequired
@@ -126,10 +139,9 @@ class OpenProjectAPIController extends Controller {
 	 * @return DataResponse
 	 */
 	public function getNotifications(): DataResponse {
-		if ($this->accessToken === '') {
-			return new DataResponse('', Http::STATUS_UNAUTHORIZED);
-		} elseif (!OpenProjectAPIService::validateURL($this->openprojectUrl)) {
-			return new DataResponse('', Http::STATUS_BAD_REQUEST);
+		$validatePreRequestResponse = $this->validatePreRequestConditions();
+		if ($validatePreRequestResponse !== null) {
+			return $validatePreRequestResponse;
 		}
 		$result = $this->openprojectAPIService->getNotifications($this->userId);
 		if (!isset($result['error'])) {
@@ -160,10 +172,9 @@ class OpenProjectAPIController extends Controller {
 		?int $fileId = null,
 		bool $isSmartPicker = false
 	): DataResponse {
-		if ($this->accessToken === '') {
-			return new DataResponse('', Http::STATUS_UNAUTHORIZED);
-		} elseif (!OpenProjectAPIService::validateURL($this->openprojectUrl)) {
-			return new DataResponse('', Http::STATUS_BAD_REQUEST);
+		$validatePreRequestResponse = $this->validatePreRequestConditions();
+		if ($validatePreRequestResponse !== null) {
+			return $validatePreRequestResponse;
 		}
 		// when the search is done through smart picker we don't want to check if the work package is linkable
 		$result = $this->openprojectAPIService->searchWorkPackage(
@@ -198,10 +209,9 @@ class OpenProjectAPIController extends Controller {
 	 * @return DataResponse
 	 */
 	public function linkWorkPackageToFile(array $values): DataResponse {
-		if ($this->accessToken === '') {
-			return new DataResponse('', Http::STATUS_UNAUTHORIZED);
-		} elseif (!OpenProjectAPIService::validateURL($this->openprojectUrl)) {
-			return new DataResponse('', Http::STATUS_BAD_REQUEST);
+		$validatePreRequestResponse = $this->validatePreRequestConditions();
+		if ($validatePreRequestResponse !== null) {
+			return $validatePreRequestResponse;
 		}
 
 		try {
@@ -227,10 +237,9 @@ class OpenProjectAPIController extends Controller {
 	 * @return DataResponse
 	 */
 	public function markNotificationAsRead(int $workpackageId) {
-		if ($this->accessToken === '') {
-			return new DataResponse('', Http::STATUS_UNAUTHORIZED);
-		} elseif (!OpenProjectAPIService::validateURL($this->openprojectUrl)) {
-			return new DataResponse('', Http::STATUS_BAD_REQUEST);
+		$validatePreRequestResponse = $this->validatePreRequestConditions();
+		if ($validatePreRequestResponse !== null) {
+			return $validatePreRequestResponse;
 		}
 		try {
 			$result = $this->openprojectAPIService->markAllNotificationsOfWorkPackageAsRead(
@@ -257,10 +266,9 @@ class OpenProjectAPIController extends Controller {
 	 * @return DataResponse
 	 */
 	public function getWorkPackageFileLinks(int $id): DataResponse {
-		if ($this->accessToken === '') {
-			return new DataResponse('', Http::STATUS_UNAUTHORIZED);
-		} elseif (!OpenProjectAPIService::validateURL($this->openprojectUrl)) {
-			return new DataResponse('', Http::STATUS_BAD_REQUEST);
+		$validatePreRequestResponse = $this->validatePreRequestConditions();
+		if ($validatePreRequestResponse !== null) {
+			return $validatePreRequestResponse;
 		}
 
 		try {
@@ -284,12 +292,10 @@ class OpenProjectAPIController extends Controller {
 	 * @return DataResponse
 	 */
 	public function deleteFileLink(int $id): DataResponse {
-		if ($this->accessToken === '') {
-			return new DataResponse('', Http::STATUS_UNAUTHORIZED);
-		} elseif (!OpenProjectAPIService::validateURL($this->openprojectUrl)) {
-			return new DataResponse('', Http::STATUS_BAD_REQUEST);
+		$validatePreRequestResponse = $this->validatePreRequestConditions();
+		if ($validatePreRequestResponse !== null) {
+			return $validatePreRequestResponse;
 		}
-
 		try {
 			$result = $this->openprojectAPIService->deleteFileLink(
 				$id,
@@ -315,10 +321,9 @@ class OpenProjectAPIController extends Controller {
 	 * @return DataResponse
 	 */
 	public function getOpenProjectWorkPackageStatus(string $id): DataResponse {
-		if ($this->accessToken === '') {
-			return new DataResponse('', Http::STATUS_UNAUTHORIZED);
-		} elseif (!OpenProjectAPIService::validateURL($this->openprojectUrl)) {
-			return new DataResponse('', Http::STATUS_BAD_REQUEST);
+		$validatePreRequestResponse = $this->validatePreRequestConditions();
+		if ($validatePreRequestResponse !== null) {
+			return $validatePreRequestResponse;
 		}
 
 		$result = $this->openprojectAPIService->getOpenProjectWorkPackageStatus(
@@ -342,10 +347,9 @@ class OpenProjectAPIController extends Controller {
 	 * @return DataResponse
 	 */
 	public function getOpenProjectWorkPackageType(string $id): DataResponse {
-		if ($this->accessToken === '') {
-			return new DataResponse('', Http::STATUS_UNAUTHORIZED);
-		} elseif (!OpenProjectAPIService::validateURL($this->openprojectUrl)) {
-			return new DataResponse('', Http::STATUS_BAD_REQUEST);
+		$validatePreRequestResponse = $this->validatePreRequestConditions();
+		if ($validatePreRequestResponse !== null) {
+			return $validatePreRequestResponse;
 		}
 
 		$result = $this->openprojectAPIService->getOpenProjectWorkPackageType(
@@ -365,10 +369,9 @@ class OpenProjectAPIController extends Controller {
 	 * @return DataResponse
 	 */
 	public function getAvailableOpenProjectProjects(?string $searchQuery = null): DataResponse {
-		if ($this->accessToken === '') {
-			return new DataResponse('', Http::STATUS_UNAUTHORIZED);
-		} elseif (!OpenProjectAPIService::validateURL($this->openprojectUrl)) {
-			return new DataResponse('', Http::STATUS_BAD_REQUEST);
+		$validatePreRequestResponse = $this->validatePreRequestConditions();
+		if ($validatePreRequestResponse !== null) {
+			return $validatePreRequestResponse;
 		}
 		try {
 			$result = $this->openprojectAPIService->getAvailableOpenProjectProjects($this->userId, $searchQuery);
@@ -417,10 +420,9 @@ class OpenProjectAPIController extends Controller {
 	 * @return DataResponse
 	 */
 	public function getOpenProjectWorkPackageForm(string $projectId, array $body): DataResponse {
-		if ($this->accessToken === '') {
-			return new DataResponse('', Http::STATUS_UNAUTHORIZED);
-		} elseif (!OpenProjectAPIService::validateURL($this->openprojectUrl)) {
-			return new DataResponse('', Http::STATUS_BAD_REQUEST);
+		$validatePreRequestResponse = $this->validatePreRequestConditions();
+		if ($validatePreRequestResponse !== null) {
+			return $validatePreRequestResponse;
 		}
 		try {
 			$result = $this->openprojectAPIService->getOpenProjectWorkPackageForm($this->userId, $projectId, $body);
@@ -439,10 +441,9 @@ class OpenProjectAPIController extends Controller {
 	 * @return DataResponse
 	 */
 	public function getAvailableAssigneesOfAProject(string $projectId): DataResponse {
-		if ($this->accessToken === '') {
-			return new DataResponse('', Http::STATUS_UNAUTHORIZED);
-		} elseif (!OpenProjectAPIService::validateURL($this->openprojectUrl)) {
-			return new DataResponse('', Http::STATUS_BAD_REQUEST);
+		$validatePreRequestResponse = $this->validatePreRequestConditions();
+		if ($validatePreRequestResponse !== null) {
+			return $validatePreRequestResponse;
 		}
 		try {
 			$result = $this->openprojectAPIService->getAvailableAssigneesOfAProject($this->userId, $projectId);
@@ -488,10 +489,9 @@ class OpenProjectAPIController extends Controller {
 	 * @return DataResponse
 	 */
 	public function createWorkPackage(array $body): DataResponse {
-		if ($this->accessToken === '') {
-			return new DataResponse('', Http::STATUS_UNAUTHORIZED);
-		} elseif (!OpenProjectAPIService::validateURL($this->openprojectUrl)) {
-			return new DataResponse('', Http::STATUS_BAD_REQUEST);
+		$validatePreRequestResponse = $this->validatePreRequestConditions();
+		if ($validatePreRequestResponse !== null) {
+			return $validatePreRequestResponse;
 		}
 		// we don't want to check if all the data in the body is set or not because
 		// that calculation will be done by the openproject api itself
@@ -517,10 +517,9 @@ class OpenProjectAPIController extends Controller {
 	 * @return DataResponse
 	 */
 	public function getOpenProjectConfiguration(): DataResponse {
-		if ($this->accessToken === '') {
-			return new DataResponse('', Http::STATUS_UNAUTHORIZED);
-		} elseif (!OpenProjectAPIService::validateURL($this->openprojectUrl)) {
-			return new DataResponse('', Http::STATUS_BAD_REQUEST);
+		$validatePreRequestResponse = $this->validatePreRequestConditions();
+		if ($validatePreRequestResponse !== null) {
+			return $validatePreRequestResponse;
 		}
 		try {
 			$result = $this->openprojectAPIService->getOpenProjectConfiguration($this->userId);
